20190518_second

+++ app/src/main/java/com/example/sensorgame2/OrientationChangeListener.java
@@ -1,5 +1,5 @@
 package com.example.sensorgame2;
 
 public interface OrientationChangeListener {
-    void onOrientationChanged(int pitch, int roll);
+    void onOrientationChanged(double pitch, double roll);
 }
